Move placeholder into separate file
