+++ packages/block-library/src/post-comments/edit.js
@@ -17,12 +17,11 @@ import {
 import { __, sprintf } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
 import { useEntityProp, store as coreStore } from '@wordpress/core-data';
-import { __experimentalUseDisabled as useDisabled } from '@wordpress/compose';
 
 /**
  * Internal dependencies
  */
-import CommentsForm from '../post-comments-form/form';
+import Placeholder from './placeholder';
 
 export default function PostCommentsEdit( {
 	attributes: { textAlign },
@@ -30,9 +29,6 @@ export default function PostCommentsEdit( {
 	context: { postType, postId },
 	clientId,
 } ) {
-	let [ postTitle ] = useEntityProp( 'postType', postType, 'title', postId );
-	postTitle = postTitle || __( 'Post Title' );
-
 	const [ commentStatus ] = useEntityProp(
 		'postType',
 		postType,
@@ -40,7 +36,7 @@ export default function PostCommentsEdit( {
 		postId
 	);
 
-	const { avatarURL, defaultCommentStatus } = useSelect(
+	const { defaultCommentStatus } = useSelect(
 		( select ) =>
 			select( blockEditorStore ).getSettings()
 				.__experimentalDiscussionSettings
@@ -94,8 +90,6 @@ export default function PostCommentsEdit( {
 		} ),
 	} );
 
-	const disabledRef = useDisabled();
-
 	if ( innerBlocks.length > 0 ) {
 		return <InnerBlocks />;
 	}
@@ -115,109 +109,7 @@ export default function PostCommentsEdit( {
 				<Warning>{ warning }</Warning>
 
 				{ showPlacholder && (
-					<div
-						className="wp-block-post-comments__placeholder"
-						ref={ disabledRef }
-					>
-						<h3>
-							{ __( 'One response to' ) } “{ postTitle }”
-						</h3>
-
-						<div className="navigation">
-							<div className="alignleft">
-								<a href="#top">« { __( 'Older Comments' ) }</a>
-							</div>
-							<div className="alignright">
-								<a href="#top">{ __( 'Newer Comments' ) } »</a>
-							</div>
-						</div>
-
-						<ol className="commentlist">
-							<li className="comment even thread-even depth-1">
-								<article className="comment-body">
-									<footer className="comment-meta">
-										<div className="comment-author vcard">
-											<img
-												alt="Commenter Avatar"
-												src={ avatarURL }
-												className="avatar avatar-32 photo"
-												height="32"
-												width="32"
-												loading="lazy"
-											/>
-											<b className="fn">
-												<a href="#top" className="url">
-													{ __(
-														'A WordPress Commenter'
-													) }
-												</a>
-											</b>{ ' ' }
-											<span className="says">
-												{ __( 'says' ) }:
-											</span>
-										</div>
-
-										<div className="comment-metadata">
-											<a href="#top">
-												<time dateTime="2000-01-01T00:00:00+00:00">
-													{ __(
-														'January 1, 2000 at 00:00 am'
-													) }
-												</time>
-											</a>{ ' ' }
-											<span className="edit-link">
-												<a
-													className="comment-edit-link"
-													href="#top"
-												>
-													{ __( 'Edit' ) }
-												</a>
-											</span>
-										</div>
-									</footer>
-
-									<div className="comment-content">
-										<p>
-											{ __( 'Hi, this is a comment.' ) }
-											<br />
-											{ __(
-												'To get started with moderating, editing, and deleting comments, please visit the Comments screen in the dashboard.'
-											) }
-											<br />
-											{ __(
-												'Commenter avatars come from'
-											) }{ ' ' }
-											<a href="https://gravatar.com/">
-												Gravatar
-											</a>
-											.
-										</p>
-									</div>
-
-									<div className="reply">
-										<a
-											className="comment-reply-link"
-											href="#top"
-											aria-label="Reply to A WordPress Commenter"
-										>
-											{ __( 'Reply' ) }
-										</a>
-									</div>
-								</article>
-							</li>
-						</ol>
-
-						<div className="navigation">
-							<div className="alignleft">
-								<a href="#top">« { __( 'Older Comments' ) }</a>
-							</div>
-							<div className="alignright">
-								<a href="#top">{ __( 'Newer Comments' ) } »</a>
-							</div>
-						</div>
-
-						<CommentsForm />
-					</div>
+					<Placeholder postId={ postId } postType={ postType } />
 				) }
 			</div>
 		</>
